add panels menu make all panels self-contained make all panels closable
