+++ packages/editor/src/components/EditorContainer.tsx
@@ -23,13 +23,12 @@ All portions of the code written by the Ethereal Engine team are Copyright © 20
 Ethereal Engine. All Rights Reserved.
 */
 
-import { DockLayout, DockMode, LayoutData, TabData } from 'rc-dock'
+import { DockLayout, DockMode, LayoutData, PanelData, TabData } from 'rc-dock'
 
 import 'rc-dock/dist/rc-dock.css'
 
 import React, { useEffect, useRef } from 'react'
 import { useHotkeys } from 'react-hotkeys-hook'
-import { useTranslation } from 'react-i18next'
 
 import { RouterState } from '@etherealengine/client-core/src/common/services/RouterService'
 import multiLogger from '@etherealengine/engine/src/common/functions/logger'
@@ -37,23 +36,16 @@ import { Engine } from '@etherealengine/engine/src/ecs/classes/Engine'
 import { EngineState } from '@etherealengine/engine/src/ecs/classes/EngineState'
 import { getMutableState, getState, useHookstate } from '@etherealengine/hyperflux'
 
-import Inventory2Icon from '@mui/icons-material/Inventory2'
 import Dialog from '@mui/material/Dialog'
 
 import { SceneState } from '@etherealengine/engine/src/ecs/classes/Scene'
 import { useQuery } from '@etherealengine/engine/src/ecs/functions/ComponentFunctions'
 import { SceneAssetPendingTagComponent } from '@etherealengine/engine/src/scene/components/SceneAssetPendingTagComponent'
 import { SceneID } from '@etherealengine/engine/src/schemas/projects/scene.schema'
-import { LocalTransformComponent } from '@etherealengine/engine/src/transform/components/TransformComponent'
 import CircularProgress from '@etherealengine/ui/src/primitives/mui/CircularProgress'
 import { t } from 'i18next'
-import { useDrop } from 'react-dnd'
-import { Vector2, Vector3 } from 'three'
-import { ItemTypes } from '../constants/AssetTypes'
-import { EditorControlFunctions } from '../functions/EditorControlFunctions'
 import { inputFileWithAddToScene } from '../functions/assetFunctions'
-import { createNewScene, getScene, saveScene } from '../functions/sceneFunctions'
-import { getCursorSpawnPosition } from '../functions/screenSpaceFunctions'
+import { getScene, onNewScene, saveScene } from '../functions/sceneFunctions'
 import { takeScreenshot } from '../functions/takeScreenshot'
 import { uploadSceneBakeToServer } from '../functions/uploadEnvMapBake'
 import { cmdOrCtrlString } from '../functions/utils'
@@ -62,8 +54,8 @@ import { EditorHelperState } from '../services/EditorHelperState'
 import { EditorState } from '../services/EditorServices'
 import './EditorContainer.css'
 import AssetDropZone from './assets/AssetDropZone'
-import ProjectBrowserPanel from './assets/ProjectBrowserPanel'
-import ScenesPanel from './assets/ScenesPanel'
+import { ProjectBrowserPanelTab } from './assets/ProjectBrowserPanel'
+import { ScenePanelTab } from './assets/ScenesPanel'
 import { ControlText } from './controlText/ControlText'
 import { DialogState } from './dialogs/DialogState'
 import ErrorDialog from './dialogs/ErrorDialog'
@@ -72,16 +64,12 @@ import SaveNewSceneDialog from './dialogs/SaveNewSceneDialog'
 import SaveSceneDialog from './dialogs/SaveSceneDialog'
 import { DndWrapper } from './dnd/DndWrapper'
 import DragLayer from './dnd/DragLayer'
-import ElementList, { SceneElementType } from './element/ElementList'
-import GraphPanel from './graph/GraphPanel'
-import { GraphPanelTitle } from './graph/GraphPanelTitle'
-import HierarchyPanelContainer from './hierarchy/HierarchyPanelContainer'
-import { HierarchyPanelTitle } from './hierarchy/HierarchyPanelTitle'
-import { PanelDragContainer, PanelIcon, PanelTitle } from './layout/Panel'
-import MaterialLibraryPanel from './materials/MaterialLibraryPanel'
-import { MaterialLibraryPanelTitle } from './materials/MaterialLibraryPanelTitle'
-import PropertiesPanelContainer from './properties/PropertiesPanelContainer'
-import { PropertiesPanelTitle } from './properties/PropertiesPanelTitle'
+import ElementList from './element/ElementList'
+import { GraphPanelTab } from './graph/GraphPanel'
+import { HierarchyPanelTab } from './hierarchy/HierarchyPanel'
+import { MaterialLibraryPanelTab } from './materials/MaterialLibraryPanel'
+import { ViewportPanelTab } from './panels/ViewportPanel'
+import { PropertiesPanelTab } from './properties/PropertiesPanel'
 import * as styles from './styles.module.scss'
 import ToolBar from './toolbar/ToolBar'
 
@@ -98,49 +86,6 @@ export const DockContainer = ({ children, id = 'editor-dock', dividerAlpha = 0 }
   return (
     <div id={id} className="dock-container" style={dockContainerStyles as React.CSSProperties}>
       {children}
-    </div>
-  )
-}
-
-const ViewportDnD = () => {
-  const [{ isDragging, isOver }, dropRef] = useDrop({
-    accept: [ItemTypes.Component],
-    collect: (monitor) => ({
-      isDragging: monitor.getItem() !== null && monitor.canDrop(),
-      isOver: monitor.isOver()
-    }),
-    drop(item: SceneElementType, monitor) {
-      const vec3 = new Vector3()
-      getCursorSpawnPosition(monitor.getClientOffset() as Vector2, vec3)
-      EditorControlFunctions.createObjectFromSceneElement([
-        { name: item!.componentJsonID },
-        { name: LocalTransformComponent.jsonID, props: { position: vec3 } }
-      ])
-    }
-  })
-
-  return (
-    <div
-      id="viewport-panel"
-      ref={dropRef}
-      style={{
-        pointerEvents: isDragging ? 'all' : 'none',
-        border: isDragging && isOver ? '5px solid white' : 'none',
-        width: '100%',
-        height: '100%'
-      }}
-    />
-  )
-}
-const ViewPortPanelContent = () => {
-  const { t } = useTranslation()
-  const sceneName = useHookstate(getMutableState(EditorState).sceneName).value
-  return sceneName ? (
-    <ViewportDnD />
-  ) : (
-    <div className={styles.bgImageBlock}>
-      <img src="/static/etherealengine.png" alt="" />
-      <h2>{t('editor:selectSceneMsg')}</h2>
     </div>
   )
 }
@@ -178,13 +123,6 @@ const SceneLoadingProgress = () => {
   )
 }
 
-const reRouteToLoadScene = async (newSceneName: string) => {
-  const { projectName, sceneName } = getState(EditorState)
-  if (sceneName === newSceneName) return
-  if (!projectName || !newSceneName) return
-  RouterState.navigate(`/studio/${projectName}/${newSceneName}`)
-}
-
 const loadScene = async (sceneName: string) => {
   const { projectName } = getState(EditorState)
   try {
@@ -197,20 +135,6 @@ const loadScene = async (sceneName: string) => {
       return
     }
     SceneState.loadScene(`${projectName}/${sceneName}` as SceneID, project)
-  } catch (error) {
-    logger.error(error)
-  }
-}
-
-const onNewScene = async () => {
-  const { projectName } = getState(EditorState)
-  if (!projectName) return
-
-  try {
-    const sceneData = await createNewScene(projectName)
-    if (!sceneData) return
-
-    reRouteToLoadScene(sceneData.name)
   } catch (error) {
     logger.error(error)
   }
@@ -390,6 +314,8 @@ const generateToolbarMenu = () => {
 
 const toolbarMenu = generateToolbarMenu()
 
+//const defaultLayout: LayoutData = useHookstate(getMutableState(EditorState).panelLayout).value
+
 const defaultLayout: LayoutData = {
   dockbox: {
     mode: 'horizontal' as DockMode,
@@ -399,28 +325,7 @@ const defaultLayout: LayoutData = {
         size: 2,
         children: [
           {
-            tabs: [
-              {
-                id: 'scenePanel',
-                title: (
-                  <PanelDragContainer>
-                    <PanelIcon as={Inventory2Icon} size={12} />
-                    <PanelTitle>Scenes</PanelTitle>
-                  </PanelDragContainer>
-                ),
-                content: <ScenesPanel newScene={onNewScene} loadScene={reRouteToLoadScene} />
-              },
-              {
-                id: 'filesPanel',
-                title: (
-                  <PanelDragContainer>
-                    <PanelIcon as={Inventory2Icon} size={12} />
-                    <PanelTitle>Files</PanelTitle>
-                  </PanelDragContainer>
-                ),
-                content: <ProjectBrowserPanel />
-              }
-            ]
+            tabs: [ScenePanelTab, ProjectBrowserPanelTab]
           }
         ]
       },
@@ -430,13 +335,7 @@ const defaultLayout: LayoutData = {
         children: [
           {
             id: '+5',
-            tabs: [
-              {
-                id: 'viewPanel',
-                title: 'Viewport',
-                content: <ViewPortPanelContent />
-              }
-            ],
+            tabs: [ViewportPanelTab],
             size: 1
           }
         ]
@@ -446,38 +345,26 @@ const defaultLayout: LayoutData = {
         size: 2,
         children: [
           {
-            tabs: [
-              {
-                id: 'hierarchyPanel',
-                title: <HierarchyPanelTitle />,
-                content: <HierarchyPanelContainer />
-              },
-              {
-                id: 'materialLibraryPanel',
-                title: <MaterialLibraryPanelTitle />,
-                content: <MaterialLibraryPanel />
-              }
-            ]
+            tabs: [HierarchyPanelTab, MaterialLibraryPanelTab]
           },
           {
-            tabs: [
-              {
-                id: 'propertiesPanel',
-                title: <PropertiesPanelTitle />,
-                content: <PropertiesPanelContainer />
-              },
-              {
-                id: 'graphPanel',
-                title: <GraphPanelTitle />,
-                content: <GraphPanel />
-              }
-            ]
+            tabs: [PropertiesPanelTab, GraphPanelTab]
           }
         ]
       }
     ]
   }
 }
+
+const panels = [
+  HierarchyPanelTab,
+  PropertiesPanelTab,
+  GraphPanelTab,
+  MaterialLibraryPanelTab,
+  ViewportPanelTab,
+  ProjectBrowserPanelTab,
+  ScenePanelTab
+]
 
 /**
  * EditorContainer class used for creating container for Editor
@@ -493,6 +380,30 @@ const EditorContainer = () => {
 
   const dialogComponent = useHookstate(getMutableState(DialogState).dialog).value
   const dockPanelRef = useRef<DockLayout>(null)
+
+  const panelMenu = panels.map((panel) => {
+    return {
+      name: panel.title,
+      action: () => {
+        const currentLayout = dockPanelRef?.current?.getLayout()
+        if (!currentLayout) return
+        let panelId = panel.id!
+        while (dockPanelRef.current!.find(panelId)) {
+          if (/\d+$/.test(panelId)) {
+            panelId = panelId.replace(/\d+$/, (match) => {
+              return (parseInt(match) + 1).toString()
+            })
+          } else {
+            panelId += '1'
+          }
+        }
+        panel.id = panelId
+        const firstPanel = currentLayout.dockbox.children[0] as PanelData
+        firstPanel.tabs.push(panel)
+        dockPanelRef?.current?.loadLayout(currentLayout)
+      }
+    }
+  })
 
   useHotkeys(`${cmdOrCtrlString}+s`, () => onSaveScene() as any)
 
@@ -539,7 +450,7 @@ const EditorContainer = () => {
       >
         <DndWrapper id="editor-container">
           <DragLayer />
-          <ToolBar menu={toolbarMenu} />
+          <ToolBar menu={toolbarMenu} panels={panelMenu} />
           <ElementList />
           <ControlText />
           {sceneLoading && <SceneLoadingProgress />}
