+++ packages/editor/src/components/assets/ProjectBrowserPanel.tsx
@@ -23,7 +23,8 @@ All portions of the code written by the Ethereal Engine team are Copyright © 20
 Ethereal Engine. All Rights Reserved.
 */
 
-import DockLayout, { DockMode } from 'rc-dock'
+import Inventory2Icon from '@mui/icons-material/Inventory2'
+import DockLayout, { DockMode, TabData } from 'rc-dock'
 import React from 'react'
 
 import { AssetSelectionChangePropsType, AssetsPreviewPanel } from './AssetsPreviewPanel'
@@ -32,6 +33,7 @@ import FileBrowserContentPanel from './FileBrowserContentPanel'
 import { getMutableState, useHookstate } from '@etherealengine/hyperflux'
 import { EditorState } from '../../services/EditorServices'
 import { DockContainer } from '../EditorContainer'
+import { PanelDragContainer, PanelIcon, PanelTitle } from '../layout/Panel'
 
 /**
  * ProjectBrowserPanel used to render view for Project Panel.
@@ -91,4 +93,16 @@ export default function ProjectBrowserPanel() {
       </DockContainer>
     </>
   )
+}
+
+export const ProjectBrowserPanelTab: TabData = {
+  id: 'filesPanel',
+  closable: true,
+  title: (
+    <PanelDragContainer>
+      <PanelIcon as={Inventory2Icon} size={12} />
+      <PanelTitle>Files</PanelTitle>
+    </PanelDragContainer>
+  ),
+  content: <ProjectBrowserPanel />
 }
