+++ packages/editor/src/components/assets/ScenesPanel.tsx
@@ -23,10 +23,6 @@ All portions of the code written by the Ethereal Engine team are Copyright © 20
 Ethereal Engine. All Rights Reserved.
 */
 
-import React, { useEffect, useState } from 'react'
-import { useTranslation } from 'react-i18next'
-import { CompressedTexture } from 'three'
-
 import { RouterState } from '@etherealengine/client-core/src/common/services/RouterService'
 import { SceneData } from '@etherealengine/common/src/interfaces/SceneInterface'
 import { AssetLoader } from '@etherealengine/engine/src/assets/classes/AssetLoader'
@@ -34,22 +30,30 @@ import createReadableTexture from '@etherealengine/engine/src/assets/functions/c
 import multiLogger from '@etherealengine/engine/src/common/functions/logger'
 import { EngineState } from '@etherealengine/engine/src/ecs/classes/EngineState'
 import { getMutableState, useHookstate } from '@etherealengine/hyperflux'
+import Inventory2Icon from '@mui/icons-material/Inventory2'
+import React, { useEffect, useState } from 'react'
+import { useTranslation } from 'react-i18next'
+import { CompressedTexture } from 'three'
 
 import { MoreVert } from '@mui/icons-material'
 import { ClickAwayListener, IconButton, InputBase, Menu, MenuItem, Paper } from '@mui/material'
 
 import { LoadingCircle } from '@etherealengine/client-core/src/components/LoadingCircle'
 import Typography from '@etherealengine/ui/src/primitives/mui/Typography'
-import { deleteScene, getScenes, renameScene } from '../../functions/sceneFunctions'
+import { TabData } from 'rc-dock'
+import { deleteScene, getScenes, onNewScene, reRouteToLoadScene, renameScene } from '../../functions/sceneFunctions'
 import { EditorState } from '../../services/EditorServices'
 import { DialogState } from '../dialogs/DialogState'
 import ErrorDialog from '../dialogs/ErrorDialog'
 import { Button } from '../inputs/Button'
+import { PanelDragContainer, PanelIcon, PanelTitle } from '../layout/Panel'
 import { InfoTooltip } from '../layout/Tooltip'
 import { DeleteDialog } from '../projects/DeleteDialog'
 import styles from './styles.module.scss'
 
 const logger = multiLogger.child({ component: 'editor:ScenesPanel' })
+
+const editorState = getMutableState(EditorState)
 
 /**
  * Displays the scenes that exist in the current project.
@@ -250,4 +254,16 @@ export default function ScenesPanel({ loadScene, newScene }) {
       />
     </>
   )
+}
+
+export const ScenePanelTab: TabData = {
+  id: 'scenePanel',
+  closable: true,
+  title: (
+    <PanelDragContainer>
+      <PanelIcon as={Inventory2Icon} size={12} />
+      <PanelTitle>Scenes</PanelTitle>
+    </PanelDragContainer>
+  ),
+  content: <ScenesPanel newScene={onNewScene} loadScene={reRouteToLoadScene} />
 }
