+++ packages/editor/src/components/toolbar/ToolBar.tsx
@@ -26,6 +26,7 @@ Ethereal Engine. All Rights Reserved.
 import React from 'react'
 
 import MenuIcon from '@mui/icons-material/Menu'
+import WindowIcon from '@mui/icons-material/Window'
 
 import DropDownMenu from '../dropDownMenu'
 import { EditorNavbarProfile } from '../projects/EditorNavbarProfile'
@@ -44,12 +45,14 @@ import TransformTool from './tools/TransformTool'
 
 type ToolBarProps = {
   menu?: any
+  panels?: any
 }
 
 export const ToolBar = (props: ToolBarProps) => {
   return (
     <div style={{ pointerEvents: 'auto' }} className={styles.toolbarContainer}>
       <DropDownMenu icon={MenuIcon} commands={props.menu} />
+      <DropDownMenu icon={WindowIcon} commands={props.panels} />
       <WorldInstanceConnection />
       <TransformTool />
       <TransformSpaceTool />
