+++ packages/editor/src/services/EditorServices.ts
@@ -25,6 +25,7 @@ Ethereal Engine. All Rights Reserved.
 
 import { EntityUUID } from '@etherealengine/common/src/interfaces/EntityUUID'
 import { defineState } from '@etherealengine/hyperflux'
+import { LayoutData } from 'rc-dock'
 
 export const EditorState = defineState({
   name: 'EditorState',
@@ -32,6 +33,7 @@ export const EditorState = defineState({
     projectName: null as string | null,
     sceneName: null as string | null,
     sceneModified: false,
-    lockPropertiesPanel: '' as EntityUUID
+    lockPropertiesPanel: '' as EntityUUID,
+    panelLayout: {} as LayoutData
   })
 })
